Created an entity called Address for an user
Following the model of the app, one of the information from users is their address. This address is not a complete one, because a user is already in a condo (that has a full address), therefore, just some information are needed in this address. For house condo, the street name and number. On the other hand, for apartment, the floor and number.